Reorganize directories for finding files easier
