+++ pgsui/impute/supervised/iterative_imputer_fixedparams.py
@@ -2,7 +2,6 @@
 import gc
 import math
 import os
-import shutil
 import sys
 import warnings
 
@@ -12,11 +11,6 @@ from time import time
 from typing import Optional, Union, List, Dict, Tuple, Any, Callable
 
 # Third-party imports
-## For plotting
-import matplotlib.pyplot as plt
-import seaborn as sns
-from matplotlib.backends.backend_pdf import PdfPages
-
 ## For stats and numeric operations
 import numpy as np
 import pandas as pd
@@ -38,28 +32,14 @@ from sklearn.utils import check_random_state, _safe_indexing, is_scalar_nan
 from sklearn.utils._mask import _get_mask
 from sklearn.utils.validation import FLOAT_DTYPES
 
-# Grid search imports
-from sklearn.model_selection import RandomizedSearchCV, GridSearchCV
-from sklearn.model_selection import StratifiedKFold
-
-# Genetic algorithm grid search imports
-from sklearn_genetic import GASearchCV
-from sklearn_genetic.callbacks import ConsecutiveStopping, DeltaThreshold
-from sklearn_genetic.plots import plot_fitness_evolution
-
-# from sklearn_genetic.space import Continuous, Categorical, Integer
-from sklearn_genetic.utils import logbook_to_pandas
-
 # Custom function imports
 try:
-    from . import simple_imputers
-    from .plotting import Plotting
-    from ..utils.misc import get_processor_name
-    from ..utils.misc import HiddenPrints
-    from ..utils.misc import isnotebook
+    from .. import simple_imputers
+    from ...utils.misc import get_processor_name
+    from ...utils.misc import HiddenPrints
+    from ...utils.misc import isnotebook
 except (ModuleNotFoundError, ValueError):
     from impute import simple_imputers
-    from impute.plotting import Plotting
     from utils.misc import get_processor_name
     from utils.misc import HiddenPrints
     from utils.misc import isnotebook
@@ -92,75 +72,59 @@ else:
 # transform method. I didn't need it, so I removed it
 # because it was saving thousands of fit estimator models into the object
 
-# _ImputerTripletGrid = namedtuple(
-# 	'_ImputerTripletGrid', ['feat_idx', 'neighbor_feat_idx', 'estimator'])
+# _ImputerTripletAll = namedtuple(
+# 	'_ImputerTripletAll', ['feat_idx', 'neighbor_feat_idx', 'estimator'])
 
 
-class IterativeImputerGridSearch(IterativeImputer):
+class IterativeImputerFixedParams(IterativeImputer):
     """Overridden IterativeImputer methods.
 
-    Herein, two types of grid searches (RandomizedSearchCV and GASearchCV), progress status updates, and several other improvements have been added. IterativeImputer is a multivariate imputer that estimates each feature from all the others. A strategy for imputing missing values by modeling each feature with missing values as a function of other features in a round-robin fashion.Read more in the scikit-learn User Guide for IterativeImputer. scikit-learn version added: 0.21. NOTE: This estimator is still **experimental** for now: the predictions and the API might change without any deprecation cycle. To use it, you need to explicitly import ``enable_iterative_imputer``\.
+    Herein, progress status updates, optimizations to save RAM, and several other improvements have been added. IterativeImputer is a multivariate imputer that estimates each feature from all the others. A strategy for imputing missing values by modeling each feature with missing values as a function of other features in a round-robin fashion.Read more in the scikit-learn User Guide for IterativeImputer. scikit-learn versionadded: 0.21. NOTE: This estimator is still **experimental** for now: the predictions and the API might change without any deprecation cycle. To use it, you need to explicitly import ``enable_iterative_imputer``\.
 
-    IterativeImputer is based on the R MICE (Multivariate Imputation by Chained Equationspackage) [1]_. See [2]_ for more information about multiple versus single imputations.
+        IterativeImputer is based on the R MICE (Multivariate Imputation by Chained Equationspackage) [1]_. See [2]_ for more information about multiple versus single imputations.
 
-    >>> # explicitly require this experimental feature
-    >>> from sklearn.experimental import enable_iterative_imputer
-    >>>
-    >>> # now you can import normally from sklearn.impute
-    >>> from sklearn.impute import IterativeImputer
+        >>> # explicitly require this experimental feature
+        >>> from sklearn.experimental import enable_iterative_imputer
+        >>>
+        >>> # now you can import normally from sklearn.impute
+        >>> from sklearn.impute import IterativeImputer
 
     Args:
         logfilepath (str): Path to the progress log file.
 
-        gridparams (sklearn_genetic.space object or Dict[str, Any]): The parameter distributions or values to use for the grid search.
-
         clf_kwargs (Dict[str, Any]): A dictionary with the classifier keyword arguments.
-
-        ga_kwargs (Dict[str, Any]): A dictionary with the genetic algorithm arguments.
 
         prefix (str): Prefix for output files.
 
-        estimator (estimator object, optional): The estimator to use at each step of the round-robin imputation. Defaults to BayesianRidge().
+        estimator (callable estimator object, optional): The estimator to use at each step of the round-robin imputation. If ``sample_posterior`` is True, the estimator must support ``return_std`` in its ``predict`` method. Defaults to BayesianRidge().
 
-        grid_cv (int, optional): The number of cross-validation folds to use with the grid search. CV folds will be stratified, and it will attempt to balance the genotypes in each fold. IMPORTANT: Sites with fewer than ``2 * grid_cv`` of each genotypes will be removed prior to doing the random subsetting because otherwise the imputation would fail. At least two of each gentoype are required to be present in each fold. Defaults to 5.
-
-        grid_n_jobs (int, optional): [The number of processors to use with the grid search. Set ``grid_n_jobs`` to -1 to use all available processors]. Defaults to 1.
-
-        grid_iter (int, optional): The number of iterations (for RandomizedSearchCV) or generations (for the genetic algorithm) to run with the grid search. For the genetic algorithm, an early stopping callback method is implemented that will stop if the accuracy does not improve for more than 5 consecutive generations. Defaults to 10.
-
-        clf_type (str, optional): Whether to run ``"classifier"`` or ``"regression"`` based imputation. Defaults to "classifier".
-
-        ga (bool, optional): Whether or not to use the genetic algorithm. If True, does genetic algorithm, otherwise does RandomizedSearchCV. Defaults to False.
+        clf_type (str, optional): Whether to run ```'classifier'``` or ``'regression'`` based imputation. Defaults to 'classifier'
 
         disable_progressbar (bool, optional): Whether or not to disable the tqdm progress bar. If True, disables the progress bar. If False, tqdm is used for the progress bar. This can be useful if you are running the imputation on an HPC cluster or are saving the standard output to a file. If True, progress updates will be printed to the screen every ``progress_update_percent`` iterations. Defaults to False.
 
-        progress_update_percent (int, optional) : How often to display progress updates (as a percentage) if ``disable_progressbar`` is True. If ``progress_update_percent=10``\, then it displays progress updates every 10%. Defaults to 10.
+        progress_update_percent (int, optional): How often to display progress updates (as a percentage) if ``disable_progressbar`` is True. If ``progress_update_frequency=10``\, then it displays progress updates every 10%. Defaults to 10.
 
-        pops (List[Union[int, str]]): List of population IDs of shape (n_samples,).
-
-        scoring_metric (str, optional): Scoring metric to use for the grid search. Defaults to "accuracy".
-
-        early_stop_gen (int, optional): Number of consecutive generations lacking improvement for which to implement the early stopping callback. Defaults to 5.
+        pops (List[Union[str, int]] or None): List of population IDs to be used with ImputeAlleleFreq if ``initial_strategy="populations"``\.
 
         missing_values (int or np.nan, optional): The placeholder for the missing values. All occurrences of ``missing_values`` will be imputed. For pandas dataframes with	nullable integer dtypes with missing values, ``missing_values`` should be set to ``np.nan``\, since ``pd.NA`` will be converted to ``np.nan``\. Defaults to np.nan.
 
-        Sample_posterior (bool, optional): CURRENTLY NOT SUPPORTED. Whether to sample from the (Gaussian) predictive posterior of the fitted estimator for each imputation. Estimator must support ``return_std`` in its ``predict`` method if set to ``True``\. Set to ``True`` if using ``IterativeImputer`` for multiple imputations. Defaults to False.
+        Sample_posterior (bool, optional): Whether to sample from the (Gaussian) predictive posterior of the fitted estimator for each imputation. Estimator must support ``return_std`` in its ``predict`` method if set to ``True``\. Set to ``True`` if using ``IterativeImputer`` for multiple imputations. Defaults to False.
 
-        max_iter (int, optional): Maximum number of imputation rounds to perform before returning the imputations computed during the final round. A round is a single	imputation of each feature with missing values. The stopping criterion is met once ``max(abs(X_t - X_{t-1}))/max(abs(X[known_vals])) < tol``\,	where ``X_t`` is ``X`` at iteration `t`. Note that early stopping is only applied if ``sample_posterior=False``\. Defaults to 10.
+        max_iter (int, optional): Maximum number of imputation rounds to perform before returning the imputations computed during the final round. A round is a single	imputation of each feature with missing values. The stopping criterion is met once ``max(abs(X_t - X_{t-1}))/max(abs(X[known_vals])) < tol``\,	where ``X_t`` is ``X`` at iteration ``t``\. Note that early stopping is only applied if ``sample_posterior=False``\. Defaults to 10.
 
         tol (float, optional): Tolerance of the stopping condition. Defaults to 1e-3.
 
         n_nearest_features (int, optional): Number of other features to use to estimate the missing values of each feature column. Nearness between features is measured using the absolute correlation coefficient between each feature pair (after initial imputation). To ensure coverage of features throughout the imputation process, the neighbor features are not necessarily nearest,	but are drawn with probability proportional to correlation for each	imputed target feature. Can provide significant speed-up when the number of features is huge. If ``None``\, all features will be used. Defaults to None.
 
-        initial_strategy (str, optional): Which strategy to use to initialize the missing values. Same as the ``strategy`` parameter in :class:`~sklearn.impute.SimpleImputer`	Valid values: "most_frequent", "populations", "nmf", or "phylogeny". Defaults to "populations".
+        initial_strategy (str, optional): Which strategy to use to initialize the missing values. Same as the ``strategy`` parameter in :class:`~sklearn.impute.SimpleImputer`	Valid values: "mean", "median", "most_frequent", "populations", "phylogeny", "nmf", or "constant". Defaults to "mean".
 
-        imputation_order (str, optional): The order in which the features will be imputed. Possible values: "ascending" (From features with fewest missing values to most), "descending" (From features with most missing values to fewest, "roman" (Left to right), "arabic" (Right to left),  random" (A random order for each round). Defaults to 'ascending'.
+        imputation_order (str, optional): The order in which the features will be imputed. Possible values: "ascending" (From features with fewest missing values to most), "descending" (From features with most missing values to fewest, "roman" (Left to right), "arabic" (Right to left),  random" (A random order for each round). Defaults to "ascending".
 
-        skip_complete (bool, optional): If True then features with missing values during ``transform`` that did not have any missing values during ``fit`` will be imputed with the initial imputation method only. Set to ``True`` if you have	many features with no missing values at both ``fit`` and ``transform`` time to save compute. Defaults to False.
+        skip_complete (bool, optional): If ``True`` then features with missing values during ``transform`` that did not have any missing values during ``fit`` will be imputed with the initial imputation method only. Set to ``True`` if you have	many features with no missing values at both ``fit`` and ``transform`` time to save compute. Defaults to False.
 
-        min_value (float or array-like of shape (n_features,), optional): Minimum possible imputed value. Broadcast to shape (n_features,) if scalar. If array-like, expects shape (n_features,), one min value for each feature. The default is `-np.inf`...versionchanged:: 0.23 (Added support for array-like). Defaults to -np.inf.
+        min_value (float or array-like of shape (n_features,), optional): Minimum possible imputed value. Broadcast to shape (n_features,) if scalar. If array-like, expects shape (n_features,), one min value for each feature. The default is `-np.inf`...sklearn versionchanged:: 0.23 (Added support for array-like). Defaults to -np.inf.
 
-        max_value (float or array-like of shape (n_features,), optional): Maximum possible imputed value. Broadcast to shape (n_features,) if scalar. If array-like, expects shape (n_features,), one max value for each feature..versionchanged:: 0.23 (Added support for array-like). Defaults to np.inf.
+        max_value (float or array-like of shape (n_features,), optional): Maximum possible imputed value. Broadcast to shape (n_features,) if scalar. If array-like, expects shape (n_features,), one max value for each feature..sklearn versionchanged:: 0.23 (Added support for array-like). Defaults to np.inf.
 
         verbose (int, optional): Verbosity flag, controls the debug messages that are issued as functions are evaluated. The higher, the more verbose. Can be 0, 1, or 2. Defaults to 0.
 
@@ -170,27 +134,61 @@ class IterativeImputerGridSearch(IterativeImputer):
 
         genotype_data (GenotypeData object, optional): GenotypeData object containing dictionary with keys=sampleIds and values=list of genotypes for the corresponding key. If using ``initial_strategy="phylogeny``\, then this object also needs contain the treefile and qmatrix objects. Defaults to None.
 
-        str_encodings (Dict[str, int], optional): Integer encodings used in STRUCTURE-formatted file. Should be a dictionary with keys=nucleotides and values=integer encodings. The missing data encoding should also be included. Argument is ignored if using a PHYLIP-formatted file. Defaults to {"A": 1, "C": 2, "G": 3, "T": 4, "N": -9}
+        str_encodings (dict(str: int), optional): Integer encodings used in STRUCTURE-formatted file. Should be a dictionary with keys=nucleotides and values=integer encodings. The missing data encoding should also be included. Argument is ignored if using a PHYLIP-formatted file. Defaults to {"A": 1, "C": 2, "G": 3, "T": 4, "N": -9}
 
     Attributes:
-        initial_imputer_: (:class:`~sklearn.impute.SimpleImputer`):  Imputer used to initialize the missing values.
-
-        imputation_sequence_ (List[Tuple[numpy.ndarray]]): Each tuple has ``(feat_idx, neighbor_feat_idx, estimator)``\, where ``feat_idx`` is the current feature to be imputed, ``neighbor_feat_idx`` is the array of other features used to impute the current feature, and ``estimator`` is the trained estimator used for the imputation. Length is ``self.n_features_with_missing_ *	self.n_iter_``\.
+        initial_imputer_ (sklearn.impute.SimpleImputer):  Imputer used to initialize the missing values.
 
         n_iter_ (int): Number of iteration rounds that occurred. Will be less than ``self.max_iter`` if early stopping criterion was reached.
 
         n_features_with_missing_ (int): Number of features with missing values.
 
-        indicator_ (:class:`~sklearn.impute.MissingIndicator`): Indicator used to add binary indicators for missing values ``None`` if add_indicator is False.
+        indicator_ (sklearn.impute.MissingIndicator): Indicator used to add binary indicators for missing values ``None`` if add_indicator is False.
 
         random_state_ (RandomState instance): RandomState instance that is generated either from a seed, the random number generator or by ``np.random``\.
+
+        logfilepath (str): Path to status logfile.
+
+        clf_kwargs (Dict[str, Any]): Keyword arguments for estimator.
+
+        prefix (str): Prefix for output files.
+
+        clf_type (str): Type of estimator, either "classifier" or "regressor".
+
+        disable_progressbar (bool): Whether to disable the tqdm progress bar. If True, writes status updates to file instead of tqdm progress bar.
+
+        progress_update_percent (float or None): Print feature progress update every ``progress_update_percent`` percent.
+
+        pops (List[Union[str, int]]): List of population IDs of shape (n_samples,).
+
+        estimator (estimator object): Estimator to impute data with.
+
+        sample_posterior (bool): Whether to use the sample_posterior option. This overridden class does not currently support sample_posterior.
+
+        max_iter (int): The maximum number of iterations to run.
+
+        tol (float): Convergence criteria.
+
+        n_nearest_features (int): Number of nearest features to impute target with.
+
+        initial_strategy (str): Strategy to use with SimpleImputer for training data.
+
+        imputation_order (str): Order to impute.
+
+        skip_complete (bool): Whether to skip features with no missing data.
+
+        min_value (int or float): Minimum value of imputed data.
+
+        max_value (int or float): Maximum value of imputed data.
+
+        verbose (int): Verbosity level.
 
         genotype_data (GenotypeData object): GenotypeData object.
 
         str_encodings (Dict[str, int]): Dictionary with integer encodings for converting from STRUCTURE-formatted file to IUPAC nucleotides.
 
     See Also:
-        SimpleImputer : Univariate imputation of missing values.
+            SimpleImputer : Univariate imputation of missing values.
 
     Examples:
         >>> import numpy as np
@@ -207,6 +205,8 @@ class IterativeImputerGridSearch(IterativeImputer):
     Notes:
         To support imputation in inductive mode we store each feature's estimator during the ``fit`` phase, and predict without refitting (in order) during	the ``transform`` phase. Features which contain all missing values at ``fit`` are discarded upon ``transform``\.
 
+        NOTE: Inductive mode support was removed herein.
+
     References:
         .. [1] Stef van Buuren, Karin Groothuis-Oudshoorn (2011). mice: Multivariate Imputation by Chained Equations in R. Journal of Statistical Software 45: 1-67.
 
@@ -217,27 +217,18 @@ class IterativeImputerGridSearch(IterativeImputer):
         self,
         logfilepath: str,
         clf_kwargs: Dict[str, Any],
-        ga_kwargs: Dict[str, Any],
         *,
         estimator: Callable = None,
-        gridparams: Dict[str, Any] = None,
-        prefix: str = "output",
-        grid_cv: int = 5,
-        grid_n_jobs: int = 1,
-        grid_iter: int = 10,
         clf_type: str = "classifier",
-        gridsearch_method: str = "gridsearch",
         disable_progressbar: bool = False,
         progress_update_percent: Optional[int] = None,
         pops: Optional[List[Union[str, int]]] = None,
-        scoring_metric: str = "accuracy",
-        early_stop_gen: int = 5,
-        missing_values: Union[int, np.float] = np.nan,
+        missing_values: Union[np.float, int] = np.nan,
         sample_posterior: bool = False,
         max_iter: int = 10,
         tol: float = 1e-3,
         n_nearest_features: Optional[int] = None,
-        initial_strategy: str = "populations",
+        initial_strategy: str = "mean",
         imputation_order: str = "ascending",
         skip_complete: bool = False,
         min_value: Union[np.float, int, float] = -np.inf,
@@ -247,8 +238,8 @@ class IterativeImputerGridSearch(IterativeImputer):
         add_indicator: bool = False,
         genotype_data: Optional[Any] = None,
         str_encodings: Optional[Dict[str, int]] = None,
+        prefix="imputer",
     ) -> None:
-
         super().__init__(
             estimator=estimator,
             missing_values=missing_values,
@@ -267,10 +258,12 @@ class IterativeImputerGridSearch(IterativeImputer):
         )
 
         self.logfilepath = logfilepath
-        self.gridparams = gridparams
         self.clf_kwargs = clf_kwargs
-        self.ga_kwargs = ga_kwargs
         self.prefix = prefix
+        self.clf_type = clf_type
+        self.disable_progressbar = disable_progressbar
+        self.progress_update_percent = progress_update_percent
+        self.pops = pops
         self.estimator = estimator
         self.sample_posterior = sample_posterior
         self.max_iter = max_iter
@@ -285,16 +278,6 @@ class IterativeImputerGridSearch(IterativeImputer):
         self.random_state = random_state
         self.genotype_data = genotype_data
         self.str_encodings = str_encodings
-        self.grid_cv = grid_cv
-        self.grid_n_jobs = grid_n_jobs
-        self.grid_iter = grid_iter
-        self.clf_type = clf_type
-        self.ga = ga
-        self.disable_progressbar = disable_progressbar
-        self.progress_update_percent = progress_update_percent
-        self.pops = pops
-        self.scoring_metric = scoring_metric
-        self.early_stop_gen = early_stop_gen
         self.missing_values = missing_values
 
     def _initial_imputation(
@@ -302,25 +285,28 @@ class IterativeImputerGridSearch(IterativeImputer):
     ) -> Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
         """Perform initial imputation for input X.
 
-        Initially imputes training data using a simple imputation method.
+        Performs initial imputation on training data (neighbors).
 
         Args:
-            X (numpy.ndarray, shape (n_samples, n_features)): Input data, where ``n_samples`` is the number of samples and ``n_features`` is the number of features.
+            X (ndarray): Input data of shape (n_samples, n_features), where n_samples is the number of samples and n_features is the number of features.
 
-            cols_to_keep (numpy.ndarray, shape (n_features,)): Column indices to keep. Only used if ``initial_strategy="phylogeny"``\.
+            cols_to_keep (numpy.ndarray): Column indices of shape (n_features,) to keep. Only used if ``initial_strategy=="phylogeny"``\.
 
-            in_fit (bool, optional): Whether function is called in fit. Defaults to False.
+            in_fit (bool, optional): True if function is called in fit, otherwise False. Defaults to False.
 
         Returns:
-            Xt (numpy.ndarray, shape (n_samples, n_features): Input data, where ``n_samples`` is the number of samples and ``n_features`` is the number of features.
+            Xt (numpy.ndarray): Input data of shape (n_samples, n_features), where n_samples is the number of samples and "n_features" is the number of features.
 
-            X_filled (numpy.ndarray, shape (n_samples, n_features)): [Input data with the most recent imputations.
+            X_filled (numpy.ndarray): Input data of shape (n_samples, features) with the most recent imputations.
 
-            mask_missing_values (numpy.ndarray, shape (n_samples, n_features)): Input data's missing indicator matrix, where ``n_samples`` is the number of samples and ``n_features`` is the number of features.
+            mask_missing_values (numpy.ndarray): Input data missing indicator matrix of of shape (n_samples, n_features), where n_samples is the number of samples and "n_features" is the number of features.
 
-            X_missing_mask (numpy.ndarray, shape (n_samples, n_features)): Input data's mask matrix indicating missing datapoints, where
-                ``n_samples`` is the number of samples and ``n_features`` is the
+            X_missing_mask (numpy.ndarray): Input data mask matrix of shape (n_samples, n_features) indicating missing datapoints, where
+                n_samples is the number of samples and n_features is the
                 number of features.
+
+        Raises:
+            AttributeError: GenotypeData object must be initialized with guidetree and qmatrix if using ``initial_strategy=phylogeny``\.
         """
         if is_scalar_nan(self.missing_values):
             force_all_finite = "allow-nan"
@@ -346,13 +332,24 @@ class IterativeImputerGridSearch(IterativeImputer):
                 by_populations=True,
                 missing=-9,
                 write_output=False,
-                output_format="array",
                 verbose=False,
                 iterative_mode=True,
                 validation_mode=True,
             )
 
-            X_filled = self.initial_imputer_.imputed
+            X_filled = self.initial_imputer_.imputed.to_numpy()
+            Xt = X.copy()
+
+        elif self.initial_strategy == "nmf":
+            self.initial_imputer_ = simple_imputers.ImputeNMF(
+                gt=np.nan_to_num(X, nan=-9),
+                missing=-9,
+                write_output=False,
+                verbose=False,
+                validation_mode=True,
+            )
+
+            X_filled = self.initial_imputer_.imputed.to_numpy()
             Xt = X.copy()
 
         elif self.initial_strategy == "phylogeny":
@@ -376,27 +373,7 @@ class IterativeImputerGridSearch(IterativeImputer):
                     validation_mode=True,
                 )
 
-                X_filled = self.initial_imputer_.imputed.to_numpy()
-                valid_sites = self.initial_imputer_.valid_sites
-
-                valid_mask = np.flatnonzero(
-                    np.logical_not(np.isnan(valid_sites))
-                )
-
-                Xt = X[:, valid_mask]
-                mask_missing_values = mask_missing_values[:, valid_mask]
-
-        elif self.initial_strategy == "nmf":
-            self.initial_imputer_ = simple_imputers.ImputeNMF(
-                gt=np.nan_to_num(X, nan=-9),
-                missing=-9,
-                write_output=False,
-                verbose=False,
-                output_format="array",
-                validation_mode=True,
-            )
-
-            X_filled = self.initial_imputer_.imputed
+            X_filled = self.initial_imputer_.imputed.to_numpy()
             Xt = X.copy()
 
         else:
@@ -410,8 +387,9 @@ class IterativeImputerGridSearch(IterativeImputer):
             else:
                 X_filled = self.initial_imputer_.transform(X)
 
-            valid_sites = self.initial_imputer_.statistics_
-            valid_mask = np.flatnonzero(np.logical_not(np.isnan(valid_sites)))
+            valid_mask = np.flatnonzero(
+                np.logical_not(np.isnan(self.initial_imputer_.statistics_))
+            )
 
             Xt = X[:, valid_mask]
             mask_missing_values = mask_missing_values[:, valid_mask]
@@ -427,7 +405,7 @@ class IterativeImputerGridSearch(IterativeImputer):
         neighbor_feat_idx: np.ndarray,
         estimator: Optional[Any] = None,
         fit_mode: bool = True,
-    ) -> Tuple[np.ndarray, Optional[Any]]:
+    ) -> np.ndarray:
         """Impute a single feature from the others provided.
 
         This function predicts the missing values of one of the features using the current estimates of all the other features. The ``estimator`` must support ``return_std=True`` in its ``predict`` method for this function to work.
@@ -441,14 +419,12 @@ class IterativeImputerGridSearch(IterativeImputer):
 
             neighbor_feat_idx (numpy.ndarray): Indices of the features to be used in imputing ``feat_idx``\.
 
-            estimator (sklearn estimator object, optional): The estimator to use at this step of the round-robin imputation If ``sample_posterior`` is True, the estimator must support ``return_std`` in its ``predict`` method. If None, it will be cloned from self._estimator. Defaults to None.
+            estimator (object): The estimator to use at this step of the round-robin imputation If ``sample_posterior`` is True, the estimator must support ``return_std`` in its ``predict`` method.If None, it will be cloned from self._estimator.
 
             fit_mode (bool, optional): Whether to fit and predict with the estimator or just predict. Defaults to True.
 
         Returns:
-            X_filled (ndarray): Input data with ``X_filled[missing_row_mask, feat_idx]`` updated.
-
-            estimator (estimator with sklearn API): The fitted estimator used to impute ``X_filled[missing_row_mask, feat_idx]``\.
+            X_filled (numpy.ndarray)]: Input data with ``X_filled[missing_row_mask, feat_idx]`` updated.
         """
         if estimator is None and fit_mode is False:
             raise ValueError(
@@ -459,57 +435,8 @@ class IterativeImputerGridSearch(IterativeImputer):
         if estimator is None:
             estimator = clone(self._estimator)
 
-        # Modified code
-        cross_val = StratifiedKFold(n_splits=self.grid_cv, shuffle=False)
-
-        # Modified code
-        # If regressor
-        if self.clf_type == "regressor":
-            if self.gridsearch_method == "genetic_algorithm":
-                callback = DeltaThreshold(threshold=1e-3, metric="fitness")
-
-        else:
-            if self.gridsearch_method == "genetic_algorithm":
-                callback = ConsecutiveStopping(
-                    generations=self.early_stop_gen, metric="fitness"
-                )
-
-        # Do randomized grid search
-        if self.gridsearch_method == "randomized_gridsearch":
-            search = RandomizedSearchCV(
-                estimator,
-                param_distributions=self.gridparams,
-                n_iter=self.grid_iter,
-                scoring=self.scoring_metric,
-                n_jobs=self.grid_n_jobs,
-                cv=cross_val,
-            )
-
-        elif self.gridsearch_method == "gridsearch":
-            search = GridSearchCV(
-                estimator,
-                param_grid=self.gridparams,
-                n_iter=self.grid_iter,
-                scoring=self.scoring_metric,
-                n_jobs=self.grid_n_jobs,
-                cv=cross_val,
-            )
-
-        # Do genetic algorithm
-        else:
-            with HiddenPrints():
-                search = GASearchCV(
-                    estimator=estimator,
-                    cv=cross_val,
-                    scoring=self.scoring_metric,
-                    generations=self.grid_iter,
-                    param_grid=self.gridparams,
-                    n_jobs=self.grid_n_jobs,
-                    verbose=False,
-                    **self.ga_kwargs,
-                )
-
         missing_row_mask = mask_missing_values[:, feat_idx]
+
         if fit_mode:
             X_train = _safe_indexing(
                 X_filled[:, neighbor_feat_idx], ~missing_row_mask
@@ -517,49 +444,26 @@ class IterativeImputerGridSearch(IterativeImputer):
 
             y_train = _safe_indexing(X_filled[:, feat_idx], ~missing_row_mask)
 
-            if self.gridsearch_method == "genetic_algorithm":
-                with HiddenPrints():
-                    search.fit(X_train, y_train, callbacks=callback)
-
-            else:
-                search.fit(X_train, y_train)
+            estimator.fit(X_train, y_train)
 
         # if no missing values, don't predict
         if np.sum(missing_row_mask) == 0:
-            return X_filled, None
+            return X_filled
 
         # get posterior samples if there is at least one missing value
         X_test = _safe_indexing(
             X_filled[:, neighbor_feat_idx], missing_row_mask
         )
 
-        # Currently un-tested with grid search
         if self.sample_posterior:
-            mus, sigmas = search.predict(X_test, return_std=True)
-            imputed_values = np.zeros(mus.shape, dtype=X_filled.dtype)
-            # two types of problems: (1) non-positive sigmas
-            # (2) mus outside legal range of min_value and max_value
-            # (results in inf sample)
-            positive_sigmas = sigmas > 0
-            imputed_values[~positive_sigmas] = mus[~positive_sigmas]
-            mus_too_low = mus < self._min_value[feat_idx]
-            imputed_values[mus_too_low] = self._min_value[feat_idx]
-            mus_too_high = mus > self._max_value[feat_idx]
-            imputed_values[mus_too_high] = self._max_value[feat_idx]
-            # the rest can be sampled without statistical issues
-            inrange_mask = positive_sigmas & ~mus_too_low & ~mus_too_high
-            mus = mus[inrange_mask]
-            sigmas = sigmas[inrange_mask]
-            a = (self._min_value[feat_idx] - mus) / sigmas
-            b = (self._max_value[feat_idx] - mus) / sigmas
-
-            truncated_normal = stats.truncnorm(a=a, b=b, loc=mus, scale=sigmas)
-            imputed_values[inrange_mask] = truncated_normal.rvs(
-                random_state=self.random_state_
+            sys.exit(
+                "sample_posterior is not currently supported. "
+                "Please set sample_posterior to False"
             )
 
         else:
-            imputed_values = search.predict(X_test)
+            imputed_values = estimator.predict(X_test)
+
             imputed_values = np.clip(
                 imputed_values,
                 self._min_value[feat_idx],
@@ -569,13 +473,14 @@ class IterativeImputerGridSearch(IterativeImputer):
         # update the feature
         X_filled[missing_row_mask, feat_idx] = imputed_values
 
+        del estimator
         del X_train
         del y_train
         del X_test
         del imputed_values
         gc.collect()
 
-        return X_filled, search
+        return X_filled
 
     @ignore_warnings(category=UserWarning)
     def fit_transform(
@@ -583,24 +488,24 @@ class IterativeImputerGridSearch(IterativeImputer):
         X: np.ndarray,
         valid_cols: Optional[np.ndarray] = None,
         y: None = None,
-    ) -> Tuple[np.ndarray, Optional[List[Any]], Optional[List[Any]]]:
+    ) -> np.ndarray:
         """Fits the imputer on X and return the transformed X.
-
-        The basic functionality is to get the nearest neightbors for each feature (column) and loop through all the features and their correlated neighbors to predict missing values.
-
-        Functionality has been added to perform grid searches using two methods (genetic algorithm and RandomSearchCV). It also makes several useful plots if using the genetic algorithm, and a tqdm progress bar and status updates have been added.
 
         Args:
             X (array-like, shape (n_samples, n_features)): Input data, where "n_samples" is the number of samples and "n_features" is the number of features.
 
             valid_cols (numpy.ndarray, optional): Array with column indices to keep. Defaults to None.
 
-            y (None, optional): Ignored. Here for compatibility with other sklearn classes. Defaults to None.
+            y (None): Ignored. Here for compatibility with other sklearn classes.
 
         Returns:
-            array-like, shape (n_samples, n_features)): The imputed input data.
-            List[Union[str, int, float]]: List of parameter settings found.
-            List[Union[int, float]]: List of scores.
+            Xt (array-like, shape (n_samples, n_features)): The imputed input data.
+
+        Raises:
+            ValueError: "max_iter" must be a positive integer.
+            ValueError: "tol" should be non-negative float.
+            ValueError: One or more features has min_value >= max_value.
+            ConvergenceWarning: Early stopping criterion not reached.
         """
         self.random_state_ = getattr(
             self, "random_state_", check_random_state(self.random_state)
@@ -627,6 +532,10 @@ class IterativeImputerGridSearch(IterativeImputer):
 
         self.initial_imputer_ = None
 
+        # X is the input data subset to only valid features (not nan)
+        # Xt is the data imputed with SimpleImputer
+        # mask_missing_values is the missing indicator matrix
+        # complete_mask is the input data's mask matrix
         X, Xt, mask_missing_values, complete_mask = self._initial_imputation(
             X, valid_cols, in_fit=True
         )
@@ -638,23 +547,15 @@ class IterativeImputerGridSearch(IterativeImputer):
 
         if self.max_iter == 0 or np.all(mask_missing_values):
             self.n_iter_ = 0
-            return (
-                super(IterativeImputer, self)._concatenate_indicator(
-                    Xt, X_indicator
-                ),
-                None,
-                None,
+            return super(IterativeImputer, self)._concatenate_indicator(
+                Xt, X_indicator
             )
 
         # Edge case: a single feature. We return the initial ...
         if Xt.shape[1] == 1:
             self.n_iter_ = 0
-            return (
-                super(IterativeImputer, self)._concatenate_indicator(
-                    Xt, X_indicator
-                ),
-                None,
-                None,
+            return super(IterativeImputer, self)._concatenate_indicator(
+                Xt, X_indicator
             )
 
         self._min_value = self._validate_limit(
@@ -692,59 +593,36 @@ class IterativeImputerGridSearch(IterativeImputer):
             Xt_previous = Xt.copy()
             normalized_tol = self.tol * np.max(np.abs(X[~mask_missing_values]))
 
-        params_list = list()
-        score_list = list()
-        iter_list = list()
-
-        if self.gridsearch_method == "genetic_algorithm":
-            sns.set_style("white")
-
         total_iter = self.max_iter
 
-        #######################################
-        ### Iterations
-        #######################################
+        ###########################################
+        ### Iteration Start
+        ###########################################
         for self.n_iter_ in progressbar(
             range(1, total_iter + 1),
             desc="Iteration: ",
             disable=self.disable_progressbar,
         ):
 
-            if self.gridsearch_method == "genetic_algorithm":
-                iter_list.append(self.n_iter_)
-
-                pp_oneline = PdfPages(
-                    f".score_traces_separate_{self.n_iter_}.pdf"
-                )
-
-                pp_lines = PdfPages(
-                    f".score_traces_combined_{self.n_iter_}.pdf"
-                )
-
-                pp_space = PdfPages(f".search_space_{self.n_iter_}.pdf")
-
             if self.imputation_order == "random":
                 ordered_idx = self._get_ordered_idx(mask_missing_values)
-
-            # Reset lists for current iteration
-            params_list.clear()
-            score_list.clear()
-            searches = list()
 
             if self.disable_progressbar:
                 with open(self.logfilepath, "a") as fout:
                     # Redirect to progress logfile
                     with redirect_stdout(fout):
                         print(
-                            f"Iteration Progress: {self.n_iter_}/{self.max_iter} ({int((self.n_iter_ / total_iter) * 100)}%)"
+                            f"Iteration Progress: "
+                            f"{self.n_iter_}/{self.max_iter} "
+                            f"({int((self.n_iter_ / total_iter) * 100)}%)"
                         )
 
                 if self.progress_update_percent is not None:
                     print_perc_interval = self.progress_update_percent
 
-            ########################################
-            ### Features
-            ########################################
+            ##########################
+            ### Feature Start
+            ##########################
             for i, feat_idx in enumerate(
                 progressbar(
                     ordered_idx,
@@ -760,7 +638,7 @@ class IterativeImputerGridSearch(IterativeImputer):
                     n_features, feat_idx, abs_corr_mat
                 )
 
-                Xt, search = self._impute_one_feature(
+                Xt = self._impute_one_feature(
                     Xt,
                     mask_missing_values,
                     feat_idx,
@@ -769,50 +647,17 @@ class IterativeImputerGridSearch(IterativeImputer):
                     fit_mode=True,
                 )
 
-                searches.append(search)
-
                 # NOTE: The below source code has been commented out to save
                 # RAM. estimator_triplet object contains numerous fit estimators
                 # that demand a lot of resources. It is primarily used for the
                 # transform function, which is not needed in this application.
 
-                # estimator_triplet = _ImputerTripletGrid(feat_idx,
-                # 									neighbor_feat_idx,
-                # 									estimator)
+                # estimator_triplet = _ImputerTripletAll(
+                # 	feat_idx,
+                # 	neighbor_feat_idx,
+                # 	estimator)
 
                 # self.imputation_sequence_.append(estimator_triplet)
-
-                if search is not None:
-                    # There was missing data in the feature
-                    params_list.append(search.best_params_)
-                    score_list.append(search.best_score_)
-
-                    if self.gridsearch_method == "genetic_algorithm":
-                        plt.cla()
-                        plt.clf()
-                        plt.close()
-
-                        plot_fitness_evolution(search)
-                        pp_oneline.savefig(bbox_inches="tight")
-                        plt.cla()
-                        plt.clf()
-                        plt.close()
-
-                        Plotting.plot_search_space(search)
-                        pp_space.savefig(bbox_inches="tight")
-                        plt.cla()
-                        plt.clf()
-                        plt.close()
-
-                else:
-                    # Search is None
-                    # Thus, there was no missing data in the given feature
-                    tmp_dict = dict()
-                    for k in self.gridparams.keys():
-                        tmp_dict[k] = -9
-                    params_list.append(tmp_dict)
-
-                    score_list.append(-9)
 
                 # Only print feature updates at each progress_update_percent
                 # interval
@@ -833,9 +678,8 @@ class IterativeImputerGridSearch(IterativeImputer):
                                     f"%)"
                                 )
 
-                            if i == len(ordered_idx):
-                                with redirect_stdout(fout):
-                                    print("")
+                                if i == len(ordered_idx):
+                                    print("\n", end="")
 
                         while print_perc_interval <= current_perc:
                             print_perc_interval += self.progress_update_percent
@@ -851,6 +695,7 @@ class IterativeImputerGridSearch(IterativeImputer):
                 inf_norm = np.linalg.norm(
                     Xt - Xt_previous, ord=np.inf, axis=None
                 )
+
                 if self.verbose > 0:
                     print(
                         f"[IterativeImputer] Change: {inf_norm}, "
@@ -873,44 +718,8 @@ class IterativeImputerGridSearch(IterativeImputer):
                             "reached."
                         )
 
-                    if self.gridsearch_method == "genetic_algorithm":
-                        pp_oneline.close()
-                        pp_space.close()
-
-                        plt.cla()
-                        plt.clf()
-                        plt.close()
-                        for iter_search in searches:
-                            if iter_search is not None:
-                                plot_fitness_evolution(iter_search)
-
-                        pp_lines.savefig(bbox_inches="tight")
-
-                        plt.cla()
-                        plt.clf()
-                        plt.close()
-                        pp_lines.close()
-
                     break
                 Xt_previous = Xt.copy()
-
-            if self.gridsearch_method == "genetic_algorithm":
-                pp_oneline.close()
-                pp_space.close()
-
-                plt.cla()
-                plt.clf()
-                plt.close()
-                for iter_search in searches:
-                    if iter_search is not None:
-                        plot_fitness_evolution(iter_search)
-
-                pp_lines.savefig(bbox_inches="tight")
-
-                plt.cla()
-                plt.clf()
-                plt.close()
-                pp_lines.close()
 
         else:
             if not self.sample_posterior:
@@ -922,33 +731,6 @@ class IterativeImputerGridSearch(IterativeImputer):
 
         Xt[~mask_missing_values] = X[~mask_missing_values]
 
-        if self.gridsearch_method == "genetic_algorithm":
-            # Remove all files except last iteration
-            final_iter = iter_list.pop()
-
-            [os.remove(f".score_traces_separate_{x}.pdf") for x in iter_list]
-            [os.remove(f".score_traces_combined_{x}.pdf") for x in iter_list]
-            [os.remove(f".search_space_{x}.pdf") for x in iter_list]
-
-            shutil.move(
-                f".score_traces_separate_{final_iter}.pdf",
-                f"{self.prefix}_score_traces_separate.pdf",
-            )
-
-            shutil.move(
-                f".score_traces_combined_{final_iter}.pdf",
-                f"{self.prefix}_score_traces_combined.pdf",
-            )
-
-            shutil.move(
-                f".search_space_{final_iter}.pdf",
-                f"{self.prefix}_search_space.pdf",
-            )
-
-        return (
-            super(IterativeImputer, self)._concatenate_indicator(
-                Xt, X_indicator
-            ),
-            params_list,
-            score_list,
+        return super(IterativeImputer, self)._concatenate_indicator(
+            Xt, X_indicator
         )
